💄 cosmetic(VSecM): rename busywait to background
busywait started to do more than waiting, so renamed it to something
that clarifies its purpose.
