+++ app/sentinel/background/initialization/token.go
